backend/requirements.txt
aioboto3==12.2.0
aiobotocore==2.9.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
boto3==1.33.13
botocore==1.33.13
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
dataclasses-json==0.6.4
distro==1.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpsx==0.0.1
httptools==0.6.1
httpx==0.26.0
idna==3.6
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-openai==0.0.5
langsmith==0.0.86
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.11.0
packaging==23.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
